app-store-google/requirements.txt
# Google Play Console API requirements
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0